requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
pydantic[email]==2.9.2
motor==3.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==12.0.0
opencv-python==4.8.1.78
torch==2.6.0
transformers==4.35.2
diffusers==0.24.0
segment-anything==1.0
deepface==0.0.79
celery==5.3.4
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
rembg==2.0.67  # Latest for background removal